refactor pins plugin
