+++ src/plugins/pins/views/pins.vue
@@ -13,22 +13,10 @@
           <span>{{ $t('plugins.pins.title') }}</span>
         </h5>
 
-        <div
-          v-if="store.isPINS && activeTab === 'software' && availableUpdatePackages.length > 0"
-          class="text-center mb-4"
-        >
-          <button
-            @click="showUpdatesModal = true"
-            class="text-amber-300 hover:text-amber-200 underline underline-offset-4 transition-colors"
-          >
-            {{ $t('plugins.pins.updatesAvailable', { count: availableUpdatePackages.length }) }}
-          </button>
-        </div>
-
         <!-- Control Panel -->
         <div v-if="store.isPINS" class="flex flex-col space-y-6 animate-fade-in-up">
           <template v-if="activeTab === 'network'">
-            <PinsWifiCard
+            <PinsNetworkTab
               :stationary-mode="stationaryMode"
               :is-scanning="isScanning"
               :wifi-list="wifiList"
@@ -46,6 +34,8 @@
               :hotspot-loading="isHotspotLoading"
               :hotspot-saving="isHotspotSaving"
               :disabled="status === 'Running'"
+              :dhcp-clients="dhcpClients"
+              :dhcp-clients-loading="isDhcpClientsLoading"
               @toggle-stationary="handleStationaryToggle"
               @scan-wifi="scanWifi"
               @refresh-interfaces="loadWifiInterfaceConfig"
@@ -61,112 +51,37 @@
               @update:hotspot-password="hotspotPassword = $event"
               @load-hotspot="loadHotspotPasswordConfig"
               @save-hotspot="saveHotspotPassword"
-            />
-
-            <PinsDhcpClientsCard
-              v-if="!stationaryMode"
-              :clients="dhcpClients"
-              :loading="isDhcpClientsLoading"
-              @refresh="loadDhcpClients"
+              @refresh-dhcp="loadDhcpClients"
             />
           </template>
 
           <template v-if="activeTab === 'services'">
-            <PinsSambaCard
-              :enabled="sambaEnabled"
+            <PinsServicesTab
+              :samba-enabled="sambaEnabled"
+              :phd2-enabled="phd2Enabled"
+              :phd2-running="phd2Running"
+              :device-time="deviceTime"
+              :time-sync-enabled="pinsStore.timeSyncEnabled"
+              :suppress-time-warning="pinsStore.suppressTimeWarning"
               :disabled="status === 'Running'"
-              @toggle="handleSambaToggle"
+              @toggle-samba="handleSambaToggle"
+              @toggle-phd2="handlePhd2Toggle"
+              @toggle-time-sync="handleTimeSyncToggle"
+              @manual-time-sync="manualTimeSync"
+              @reenable-time-warning="pinsStore.setSuppressTimeWarning(false)"
             />
-
-            <PinsPhd2Card
-              :enabled="phd2Enabled"
-              :running="phd2Running"
-              :disabled="status === 'Running'"
-              @toggle="handlePhd2Toggle"
-            />
-
-            <!-- System Time Card -->
-            <div
-              class="border border-gray-700 rounded-lg bg-gray-800 shadow-xl p-6 relative overflow-hidden flex flex-row items-center justify-between"
-            >
-              <div class="absolute top-0 right-20 p-4 opacity-10 pointer-events-none">
-                <svg
-                  class="w-16 h-16 text-white"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    stroke-width="2"
-                    d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"
-                  />
-                </svg>
-              </div>
-              <div class="relative z-10">
-                <h3 class="text-xl font-bold text-white mb-1">
-                  {{ $t('plugins.pins.systemTime') }}
-                </h3>
-                <p class="text-gray-400 text-sm" v-if="deviceTime">
-                  {{ $t('plugins.pins.deviceTime') }}:
-                  {{ new Date(deviceTime * 1000).toLocaleString() }}
-                </p>
-                <p class="text-gray-400 text-sm" v-else>
-                  {{ $t('plugins.pins.loadingTime') }}
-                </p>
-                <button
-                  v-if="pinsStore.suppressTimeWarning"
-                  @click="pinsStore.setSuppressTimeWarning(false)"
-                  class="text-xs text-yellow-400 hover:text-yellow-300 underline mt-1 text-left"
-                >
-                  {{ $t('plugins.pins.timeWarning.reenable') }}
-                </button>
-              </div>
-              <div class="relative z-10 flex flex-col items-end gap-2">
-                <div class="flex items-center gap-2">
-                  <span class="text-xs text-gray-400 uppercase font-bold">{{
-                    $t('plugins.pins.autoSync')
-                  }}</span>
-                  <toggleButton
-                    :status-value="pinsStore.timeSyncEnabled"
-                    @update:status-value="handleTimeSyncToggle"
-                    :disabled="status === 'Running'"
-                  />
-                </div>
-                <button
-                  @click="manualTimeSync"
-                  class="flex items-center gap-2 px-4 py-2 bg-gradient-to-r from-blue-600 to-blue-700 hover:from-blue-500 hover:to-blue-600 text-white font-bold rounded-lg shadow-md shadow-blue-900/20 transition-all disabled:opacity-50 text-xs sm:text-sm"
-                  :disabled="status === 'Running'"
-                >
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    class="h-4 w-4"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      stroke-linecap="round"
-                      stroke-linejoin="round"
-                      stroke-width="2"
-                      d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
-                    />
-                  </svg>
-                  {{ $t('plugins.pins.syncNow') }}
-                </button>
-              </div>
-            </div>
           </template>
 
           <template v-if="activeTab === 'software'">
-            <PinsIndi3rdpartyCard
+            <PinsSoftwareTab
+              :available-update-count="availableUpdatePackages.length"
               :drivers="indi3rdpartyDrivers"
               :loading="isIndi3rdpartyLoading"
               :installing="isIndi3rdpartyInstalling"
               :search-query="indi3rdpartyQuery"
               :selected-asset="selectedIndi3rdpartyAsset"
               :disabled="status === 'Running'"
+              @open-updates="showUpdatesModal = true"
               @refresh="loadIndi3rdpartyDrivers"
               @search="loadIndi3rdpartyDrivers"
               @install="installIndi3rdpartyDriver"
@@ -176,14 +91,14 @@
           </template>
 
           <template v-if="activeTab === 'upgrade'">
-            <PinsUpgradeCard
+            <PinsUpgradeTab
               :status="status"
               :active-operation="activeOperation"
               :upgrade-exit-code="upgradeExitCode"
+              :logs="logs"
               @start-upgrade="startUpgrade"
+              @clear-logs="pinsStore.clearTerminalLogs()"
             />
-
-            <PinsTerminalOutput :logs="logs" @clear="pinsStore.clearTerminalLogs()" />
           </template>
 
           <Modal
@@ -319,16 +234,14 @@ import { useSettingsStore } from '@/store/settingsStore';
 import { usePinsStore } from '../store/pinsStore';
 import { apiStore } from '@/store/store';
 import axios from 'axios';
-import toggleButton from '@/components/helpers/toggleButton.vue';
 import SubNav from '@/components/SubNav.vue';
 import Modal from '@/components/helpers/Modal.vue';
-import PinsUpgradeCard from '../components/PinsUpgradeCard.vue';
-import PinsTerminalOutput from '../components/PinsTerminalOutput.vue';
-import PinsSambaCard from '../components/PinsSambaCard.vue';
-import PinsPhd2Card from '../components/PinsPhd2Card.vue';
-import PinsWifiCard from '../components/PinsWifiCard.vue';
-import PinsDhcpClientsCard from '../components/PinsDhcpClientsCard.vue';
-import PinsIndi3rdpartyCard from '../components/PinsIndi3rdpartyCard.vue';
+import PinsNetworkTab from '../components/tabs/PinsNetworkTab.vue';
+import PinsServicesTab from '../components/tabs/PinsServicesTab.vue';
+import PinsSoftwareTab from '../components/tabs/PinsSoftwareTab.vue';
+import PinsUpgradeTab from '../components/tabs/PinsUpgradeTab.vue';
+import { usePinsWifiInterfaces } from '../composables/usePinsWifiInterfaces';
+import { usePinsUpgradeTracker } from '../composables/usePinsUpgradeTracker';
 
 const { t } = useI18n();
 const settingsStore = useSettingsStore();
@@ -355,11 +268,6 @@ const dhcpClients = ref([]);
 const isDhcpClientsLoading = ref(false);
 const selectedIndi3rdpartyAsset = ref('');
 const activeTab = ref('network');
-const wifiAdapters = ref([]);
-const isWifiAdaptersLoading = ref(false);
-const isWifiInterfacesSaving = ref(false);
-const selectedClientInterface = ref('');
-const selectedHotspotInterface = ref('');
 const {
   stationaryMode,
   wifiList,
@@ -379,20 +287,6 @@ let isComponentUnmounting = false;
 
 const PORT = 8000;
 const TOKEN = 'zZDqJ3IKeFaIZqG2JIFvsxzA5E48GC2gyGVagHFZqC0OMtgoupUDZCPhQDYKm35d';
-const LAST_UPGRADE_JOB_ID_KEY = 'lastUpgradeJobId';
-const LAST_UPGRADE_JOB_RESULT_KEY = 'lastUpgradeJobResult';
-const UPGRADE_POLL_INTERVAL_MS = 3000;
-const UPGRADE_INITIAL_BACKOFF_MS = 2000;
-const UPGRADE_MAX_BACKOFF_MS = 30000;
-
-const upgradeExitCode = ref(null);
-let hasRestoredUpgradeState = false;
-let upgradePollTimer = null;
-let upgradePollBackoffMs = UPGRADE_INITIAL_BACKOFF_MS;
-let isUpgradePolling = false;
-let lastUpgradeStatus = null;
-let upgradeLatestNotFoundCount = 0;
-const UPGRADE_LATEST_NOT_FOUND_MAX_RETRIES = 2;
 
 const availableUpdatePackages = computed(() => {
   const packages = updatesCheckResult.value?.packages || [];
@@ -422,520 +316,41 @@ function appendLog(message) {
   pinsStore.appendTerminalLog(message);
 }
 
-function normalizeWifiAdapter(adapter) {
-  if (!adapter || typeof adapter !== 'object') {
-    return null;
-  }
-
-  return {
-    interface: adapter.interface || '',
-    state: adapter.state || 'unknown',
-    connection: adapter.connection ?? null,
-    role: adapter.role || 'idle',
-    mac: adapter.mac ?? null,
-    driver: adapter.driver ?? null,
-    mtu: adapter.mtu ?? null,
-  };
-}
-
-function reconcileSelectedWifiInterfaces() {
-  const availableIfaces = new Set(wifiAdapters.value.map((adapter) => adapter.interface));
-
-  if (selectedClientInterface.value && !availableIfaces.has(selectedClientInterface.value)) {
-    selectedClientInterface.value = '';
-  }
-
-  if (selectedHotspotInterface.value && !availableIfaces.has(selectedHotspotInterface.value)) {
-    selectedHotspotInterface.value = '';
-  }
-}
-
-async function loadWifiAdapters() {
-  const ip = getIp();
-  if (!ip || isWifiAdaptersLoading.value) return;
-
-  isWifiAdaptersLoading.value = true;
-  try {
-    const directAxios = axios.create({ headers: {} });
-    const response = await directAxios.get(`http://${ip}:${PORT}/wifi/adapters`, {
-      headers: {
-        Authorization: `Bearer ${TOKEN}`,
-      },
-      timeout: 10000,
-    });
-
-    const adapters = response.data?.adapters || [];
-    wifiAdapters.value = adapters.map(normalizeWifiAdapter).filter(Boolean);
-    reconcileSelectedWifiInterfaces();
-  } catch (error) {
-    console.error(error);
-    appendLog(t('plugins.pins.logs.wifiAdaptersLoadFailed', { message: error.message }));
-  } finally {
-    isWifiAdaptersLoading.value = false;
-  }
-}
-
-async function loadWifiInterfaces() {
-  const ip = getIp();
-  if (!ip) return;
-
-  try {
-    const directAxios = axios.create({ headers: {} });
-    const response = await directAxios.get(`http://${ip}:${PORT}/wifi/interfaces`, {
-      headers: {
-        Authorization: `Bearer ${TOKEN}`,
-      },
-      timeout: 10000,
-    });
-
-    const data = response.data || {};
-    selectedClientInterface.value = data.client_interface || '';
-    selectedHotspotInterface.value = data.hotspot_interface || '';
-    reconcileSelectedWifiInterfaces();
-  } catch (error) {
-    console.error(error);
-    appendLog(t('plugins.pins.logs.wifiInterfacesLoadFailed', { message: error.message }));
-  }
-}
-
-async function loadWifiInterfaceConfig() {
-  await Promise.all([loadWifiAdapters(), loadWifiInterfaces()]);
-  reconcileSelectedWifiInterfaces();
-}
-
-async function saveWifiInterfaces() {
-  if (status.value === 'Running' || isWifiInterfacesSaving.value) return;
-
-  const ip = getIp();
-  if (!ip) {
-    appendLog(t('plugins.pins.logs.noIp'));
-    return;
-  }
-
-  isWifiInterfacesSaving.value = true;
-  try {
-    const directAxios = axios.create({ headers: {} });
-    const response = await directAxios.post(
-      `http://${ip}:${PORT}/wifi/interfaces`,
-      {
-        client_interface: selectedClientInterface.value || null,
-        hotspot_interface: selectedHotspotInterface.value || null,
-      },
-      {
-        headers: {
-          Authorization: `Bearer ${TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        timeout: 10000,
-      }
-    );
-
-    const data = response.data || {};
-    selectedClientInterface.value = data.client_interface || '';
-    selectedHotspotInterface.value = data.hotspot_interface || '';
-    reconcileSelectedWifiInterfaces();
-    appendLog(
-      t('plugins.pins.logs.wifiInterfacesSaved', {
-        clientInterface: selectedClientInterface.value || 'auto',
-        hotspotInterface: selectedHotspotInterface.value || 'auto',
-      })
-    );
-    await loadWifiAdapters();
-  } catch (error) {
-    console.error(error);
-    appendLog(t('plugins.pins.logs.wifiInterfacesSaveFailed', { message: error.message }));
-
-    if (error.response) {
-      appendLog(
-        t('plugins.pins.logs.serverError', {
-          status: error.response.status,
-          data: JSON.stringify(error.response.data),
-        })
-      );
-    }
-  } finally {
-    isWifiInterfacesSaving.value = false;
-  }
-}
-
-function getStorageItem(key) {
-  try {
-    return window.localStorage.getItem(key);
-  } catch {
-    return null;
-  }
-}
-
-function setStorageItem(key, value) {
-  try {
-    window.localStorage.setItem(key, value);
-  } catch {
-    // Ignore storage write errors.
-  }
-}
-
-function removeStorageItem(key) {
-  try {
-    window.localStorage.removeItem(key);
-  } catch {
-    // Ignore storage remove errors.
-  }
-}
-
-function toFiniteNumber(value) {
-  if (value === null || value === undefined) {
-    return null;
-  }
-  const parsed = Number(value);
-  return Number.isFinite(parsed) ? parsed : null;
-}
-
-function normalizeUpgradeJob(rawJob) {
-  if (!rawJob || typeof rawJob !== 'object') {
-    return null;
-  }
-
-  return {
-    jobId: rawJob.jobId ?? null,
-    status: String(rawJob.status ?? '').toLowerCase(),
-    exitCode: rawJob.exitCode ?? rawJob.exit_code ?? null,
-    startedAt: toFiniteNumber(rawJob.startedAt),
-    finishedAt: toFiniteNumber(rawJob.finishedAt),
-    command: rawJob.command ?? null,
-  };
-}
-
-function isUpgradeTerminalStatus(statusValue) {
-  return statusValue === 'success' || statusValue === 'failed';
-}
-
-function isUpgradeInProgressStatus(statusValue) {
-  return statusValue === 'started' || statusValue === 'running';
-}
-
-function setUpgradeJobId(nextJobId) {
-  pinsStore.setCurrentJobId(nextJobId || null);
-  if (nextJobId) {
-    setStorageItem(LAST_UPGRADE_JOB_ID_KEY, String(nextJobId));
-  } else {
-    removeStorageItem(LAST_UPGRADE_JOB_ID_KEY);
-  }
-}
-
-function getStoredUpgradeJobId() {
-  return getStorageItem(LAST_UPGRADE_JOB_ID_KEY);
-}
-
-function saveUpgradeFinalResult(job) {
-  setStorageItem(LAST_UPGRADE_JOB_RESULT_KEY, JSON.stringify(job));
-}
-
-function clearStoredUpgradeFinalResult() {
-  removeStorageItem(LAST_UPGRADE_JOB_RESULT_KEY);
-}
-
-function getStoredUpgradeFinalResult() {
-  const raw = getStorageItem(LAST_UPGRADE_JOB_RESULT_KEY);
-  if (!raw) {
-    return null;
-  }
-
-  try {
-    return normalizeUpgradeJob(JSON.parse(raw));
-  } catch {
-    return null;
-  }
-}
-
-function clearUpgradePollTimer() {
-  if (upgradePollTimer) {
-    clearTimeout(upgradePollTimer);
-    upgradePollTimer = null;
-  }
-}
-
-function stopUpgradePolling() {
-  isUpgradePolling = false;
-  clearUpgradePollTimer();
-  upgradePollBackoffMs = UPGRADE_INITIAL_BACKOFF_MS;
-  upgradeLatestNotFoundCount = 0;
-}
-
-function scheduleUpgradePolling(ip, delayMs = UPGRADE_POLL_INTERVAL_MS) {
-  if (!isUpgradePolling || !ip) {
-    return;
-  }
-
-  clearUpgradePollTimer();
-  upgradePollTimer = setTimeout(() => {
-    pollUpgradeStatus(ip);
-  }, delayMs);
-}
-
-function shouldRetryUpgradePolling(error) {
-  const statusCode = error?.response?.status;
-  if (!error?.response) {
-    return true;
-  }
-  if (statusCode >= 500) {
-    return true;
-  }
-  return statusCode === 408 || statusCode === 429;
-}
-
-function applyUpgradeJobState(job, { persistFinal = true } = {}) {
-  const normalized = normalizeUpgradeJob(job);
-  if (!normalized) {
-    return false;
-  }
-
-  const previousStatus = lastUpgradeStatus;
-  if (normalized.jobId) {
-    setUpgradeJobId(normalized.jobId);
-  }
-
-  if (normalized.status && normalized.status !== previousStatus) {
-    appendLog(t('plugins.pins.logs.jobStatus', { status: normalized.status }));
-  }
-
-  if (normalized.exitCode !== null && normalized.exitCode !== undefined) {
-    upgradeExitCode.value = normalized.exitCode;
-  }
-
-  lastUpgradeStatus = normalized.status || previousStatus;
-  pinsStore.setActiveOperation('upgrade');
-
-  if (persistFinal || !isUpgradeTerminalStatus(normalized.status)) {
-    saveUpgradeFinalResult(normalized);
-  }
-
-  if (normalized.status === 'success') {
-    status.value = 'Success';
-    if (previousStatus !== 'success') {
-      appendLog(t('plugins.pins.logs.upgradeSuccess'));
-    }
-    stopUpgradePolling();
-    return true;
-  }
-
-  if (normalized.status === 'failed') {
-    status.value = 'Failed';
-    if (previousStatus !== 'failed') {
-      appendLog(
-        t('plugins.pins.logs.upgradeFailed', {
-          exitCode: normalized.exitCode ?? 'Unknown',
-        })
-      );
-    }
-    stopUpgradePolling();
-    return true;
-  }
-
-  if (isUpgradeInProgressStatus(normalized.status)) {
-    status.value = 'Running';
-    return false;
-  }
-
-  status.value = 'Running';
-  return false;
-}
-
-async function fetchUpgradeJobById(ip, id) {
-  const directAxios = axios.create({ headers: {} });
-  const response = await directAxios.get(`http://${ip}:${PORT}/jobs/${id}`, {
-    headers: {
-      Authorization: `Bearer ${TOKEN}`,
-    },
-    timeout: 8000,
-  });
-
-  return normalizeUpgradeJob(response.data);
-}
-
-async function fetchLatestUpgradeJob(ip) {
-  const directAxios = axios.create({ headers: {} });
-  const response = await directAxios.get(`http://${ip}:${PORT}/jobs/latest`, {
-    headers: {
-      Authorization: `Bearer ${TOKEN}`,
-    },
-    timeout: 8000,
-  });
-
-  return normalizeUpgradeJob(response.data);
-}
-
-function selectPreferredJob(storedFinalResult, latestJob) {
-  if (!storedFinalResult) {
-    return latestJob;
-  }
-  if (!latestJob) {
-    return storedFinalResult;
-  }
-  if (storedFinalResult.jobId === latestJob.jobId) {
-    return latestJob;
-  }
-
-  const storedStartedAt = storedFinalResult.startedAt ?? -Infinity;
-  const latestStartedAt = latestJob.startedAt ?? -Infinity;
-  return latestStartedAt >= storedStartedAt ? latestJob : storedFinalResult;
-}
-
-async function pollUpgradeStatus(ip) {
-  if (!isUpgradePolling) {
-    return;
-  }
-
-  const trackedJobId = jobId.value || getStoredUpgradeJobId();
-  if (!trackedJobId) {
-    stopUpgradePolling();
-    return;
-  }
-
-  try {
-    const currentJob = await fetchUpgradeJobById(ip, trackedJobId);
-    upgradePollBackoffMs = UPGRADE_INITIAL_BACKOFF_MS;
-    upgradeLatestNotFoundCount = 0;
-    const terminal = applyUpgradeJobState(currentJob);
-    if (!terminal) {
-      scheduleUpgradePolling(ip);
-    }
-    return;
-  } catch (error) {
-    if (error?.response?.status === 404) {
-      try {
-        const latestJob = await fetchLatestUpgradeJob(ip);
-        const storedFinal = getStoredUpgradeFinalResult();
-        const selectedJob = selectPreferredJob(storedFinal, latestJob);
-
-        if (selectedJob?.jobId && selectedJob.jobId !== trackedJobId) {
-          setUpgradeJobId(selectedJob.jobId);
-        }
-
-        upgradePollBackoffMs = UPGRADE_INITIAL_BACKOFF_MS;
-        const terminal = applyUpgradeJobState(selectedJob);
-        if (!terminal) {
-          scheduleUpgradePolling(ip);
-        }
-        return;
-      } catch (latestError) {
-        if (latestError?.response?.status === 404) {
-          upgradeLatestNotFoundCount += 1;
-          if (upgradeLatestNotFoundCount <= UPGRADE_LATEST_NOT_FOUND_MAX_RETRIES) {
-            const retryDelay = Math.min(upgradePollBackoffMs, UPGRADE_MAX_BACKOFF_MS);
-            appendLog(
-              t('plugins.pins.logs.error', {
-                message: `No jobs found yet. Retrying in ${Math.round(retryDelay / 1000)}s.`,
-              })
-            );
-            upgradePollBackoffMs = Math.min(upgradePollBackoffMs * 2, UPGRADE_MAX_BACKOFF_MS);
-            scheduleUpgradePolling(ip, retryDelay);
-            return;
-          }
-
-          appendLog(
-            t('plugins.pins.logs.error', {
-              message: 'No upgrade jobs available. Upgrade tracking stopped.',
-            })
-          );
-
-          stopUpgradePolling();
-          setUpgradeJobId(null);
-          lastUpgradeStatus = null;
-
-          // If there is no terminal result saved, return controls to idle.
-          const latestStoredResult = getStoredUpgradeFinalResult();
-          if (!latestStoredResult || !isUpgradeTerminalStatus(latestStoredResult.status)) {
-            status.value = 'Idle';
-            pinsStore.setActiveOperation(null);
-            upgradeExitCode.value = null;
-          }
-          return;
-        }
-
-        error = latestError;
-      }
-    }
-
-    if (error?.response?.status === 401) {
-      appendLog(
-        t('plugins.pins.logs.serverError', {
-          status: error.response.status,
-          data: JSON.stringify(error.response.data),
-        })
-      );
-      appendLog(
-        t('plugins.pins.logs.error', {
-          message: 'Unauthorized. Please re-authenticate or refresh your token.',
-        })
-      );
-      status.value = 'Failed';
-      stopUpgradePolling();
-      return;
-    }
-
-    if (shouldRetryUpgradePolling(error)) {
-      const retryDelay = Math.min(upgradePollBackoffMs, UPGRADE_MAX_BACKOFF_MS);
-      upgradePollBackoffMs = Math.min(upgradePollBackoffMs * 2, UPGRADE_MAX_BACKOFF_MS);
-      appendLog(
-        t('plugins.pins.logs.error', {
-          message: `Status check failed (${error.message}). Retrying in ${Math.round(
-            retryDelay / 1000
-          )}s.`,
-        })
-      );
-      scheduleUpgradePolling(ip, retryDelay);
-      return;
-    }
-
-    appendLog(t('plugins.pins.logs.statusFetchFailed', { message: error.message }));
-    status.value = 'Failed';
-    stopUpgradePolling();
-  }
-}
-
-function beginUpgradeTracking(ip, nextJobId) {
-  if (!ip || !nextJobId) {
-    return;
-  }
-
-  isUpgradePolling = true;
-  upgradePollBackoffMs = UPGRADE_INITIAL_BACKOFF_MS;
-  upgradeLatestNotFoundCount = 0;
-  setUpgradeJobId(nextJobId);
-  scheduleUpgradePolling(ip, 0);
-}
-
-function restoreUpgradeState() {
-  if (hasRestoredUpgradeState) {
-    return;
-  }
-  hasRestoredUpgradeState = true;
-
-  const ip = getIp();
-  if (!ip) {
-    return;
-  }
-
-  const storedFinalResult = getStoredUpgradeFinalResult();
-  if (storedFinalResult && isUpgradeTerminalStatus(storedFinalResult.status)) {
-    pinsStore.setActiveOperation('upgrade');
-    lastUpgradeStatus = storedFinalResult.status;
-    upgradeExitCode.value = storedFinalResult.exitCode;
-    status.value = storedFinalResult.status === 'success' ? 'Success' : 'Failed';
-  }
-
-  const trackedJobId = jobId.value || getStoredUpgradeJobId();
-  if (!trackedJobId) {
-    return;
-  }
-
-  setUpgradeJobId(trackedJobId);
-  if (!storedFinalResult || !isUpgradeTerminalStatus(storedFinalResult.status)) {
-    pinsStore.setActiveOperation('upgrade');
-    status.value = 'Running';
-    beginUpgradeTracking(ip, trackedJobId);
-  }
-}
+const {
+  wifiAdapters,
+  isWifiAdaptersLoading,
+  isWifiInterfacesSaving,
+  selectedClientInterface,
+  selectedHotspotInterface,
+  loadWifiInterfaceConfig,
+  saveWifiInterfaces,
+} = usePinsWifiInterfaces({
+  t,
+  appendLog,
+  getIp,
+  PORT,
+  TOKEN,
+  status,
+});
+
+const {
+  upgradeExitCode,
+  stopUpgradePolling,
+  restoreUpgradeState,
+  beginUpgradeTrackingFromStart,
+  handleUpgradeWsClosed,
+  resetUpgradeForNewRun,
+} = usePinsUpgradeTracker({
+  t,
+  appendLog,
+  pinsStore,
+  status,
+  jobId,
+  activeOperation,
+  getIp,
+  PORT,
+  TOKEN,
+});
 
 watch(
   () => store.isPINS,
@@ -1733,10 +1148,7 @@ async function startUpgrade() {
   status.value = 'Running';
   pinsStore.setActiveOperation('upgrade');
   pinsStore.clearTerminalLogs();
-  stopUpgradePolling();
-  clearStoredUpgradeFinalResult();
-  upgradeExitCode.value = null;
-  lastUpgradeStatus = null;
+  resetUpgradeForNewRun();
   appendLog(t('plugins.pins.logs.init', { ip }));
 
   try {
@@ -1771,18 +1183,8 @@ async function startUpgrade() {
       throw new Error('No valid Job ID returned. server response: ' + JSON.stringify(data));
     }
 
-    setUpgradeJobId(returnedJobId);
     appendLog(t('plugins.pins.logs.jobCreated', { jobId: returnedJobId }));
-
-    const normalizedResponseJob = normalizeUpgradeJob(data);
-    if (normalizedResponseJob) {
-      applyUpgradeJobState(
-        { ...normalizedResponseJob, jobId: returnedJobId },
-        { persistFinal: false }
-      );
-    }
-
-    beginUpgradeTracking(ip, returnedJobId);
+    beginUpgradeTrackingFromStart(ip, returnedJobId, data);
 
     connectWebSocket(ip, returnedJobId);
   } catch (error) {
@@ -1850,8 +1252,7 @@ function connectWebSocket(ip, id) {
       ws = null;
 
       // Upgrade status resolution uses polling with /jobs/{id} and /jobs/latest fallback.
-      if (activeOperation.value === 'upgrade' && isUpgradePolling) {
-        scheduleUpgradePolling(ip, 0);
+      if (handleUpgradeWsClosed(ip)) {
         return;
       }
